working so far how it is

+++ flake.nix
@@ -1,7 +1,5 @@
-# The python3Packages is define in the framework docs
-# https://github.com/NixOS/nixpkgs/tree/master/doc/languages-frameworks
 {
-  description = "Nix flake for a Python3 project with requirements.txt and Docker support";
+  description = "Nix flake for a Python development environment using pip and venv";
 
   inputs = {
     nixpkgs.url = "github:NixOS/nixpkgs/nixos-unstable";
@@ -18,33 +16,33 @@
       system:
       let
         pkgs = import nixpkgs { inherit system; };
-        pythonEnv = pkgs.python3Packages.buildPythonPackage {
-          pname = "my-python-project";
-          version = "0.1.0";
-          src = ./.;
-          buildInputs = [
-            pkgs.docker
-            pkgs.git
-          ];
-          propagatedBuildInputs = with pkgs.python3Packages; [
-            pip
-            setuptools
-            wheel
-          ];
-          # Install dependencies from requirements.txt
-          preBuild = ''
-            pip install -r requirements.txt
-          '';
-        };
       in
       {
         devShell = pkgs.mkShell {
-          buildInputs = [
-            pythonEnv
-            pkgs.docker # Include Docker in the environment
+          # Include Python and pip in the environment
+          packages = [
+            pkgs.python313
+            pkgs.python313Packages.pip
+                        #pkgs.python313Packages.psycopg2
+            pkgs.git
+            pkgs.docker
+            pkgs.direnv
           ];
+
+          # Create and activate a virtual environment, then install requirements
           shellHook = ''
-            echo "Welcome to the Python3 + Docker dev environment!"
+            if [ ! -d .venv ]; then
+              echo "Creating virtual environment..."
+              python -m venv .venv
+            fi
+            source .venv/bin/activate
+            pip install --upgrade pip
+            if [ -f requirements.txt ]; then
+              echo "Installing Python dependencies from requirements.txt..."
+              pip install -r requirements.txt
+            else
+              echo "No requirements.txt found."
+            fi
           '';
         };
       }
